backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
kiteconnect==5.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
redis==5.2.1
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.28.1
pandas==2.2.3
numpy==2.2.1
scipy==1.14.1
python-telegram-bot==21.9
websockets==14.1
pydantic==2.10.3
